backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
openai==1.54.0
tavily-python==0.5.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0
dashscope>=1.25.6
